expanded token models
For now student tokens for the measurement can be null. This will be changed in the future, but for now this is fine. The migrations would be a pain otherwise.

Also feel free to remover more allowed characters from the list in settings. I might not have removed all necessary ones.