added prototype of histogram to range filter
This is a first prototype of a possible histogram to the range filter. It is quite hard to get the slider to exactly match the bars of the histogram. Ideally with many measurements this is not going to be a huge issue as long as the slider and histogram somewhat correlate and the error is not too large...