Introduce new app for export (CSV and XLSX)
Addresses issue #78.

To implement the export views, a custom response class
`MeasurementExportResponse` is introduced, which in turn is a subclass
of `django.http.FileResponse`. Instead of streaming the contents of
the files on-the-fly as in the previous approach, a temporary file is
created and served. This enables the use of the `xlsxwriter` package,
allowing detailed customisation of the XLSX file and its formatting.